[refactor] : return ResponseEntity
